Add files via upload
This is the first version of the code implementation of LeNet with PReLU activation function using Keras and TensorFlow. The markdown explaining the code and the technical background of the network will be available in a future update of the NoteBook. The comments in the Notebook should suffice for now.